Restore persisted prompt and generation evidence at the chat IPC boundary
